backend/requirements.txt
# Core Flask and Extensions
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
Flask-JWT-Extended==4.6.0

# Database
psycopg2-binary==2.9.9
alembic==1.13.1
SQLAlchemy==2.0.23  # Explicitly specify SQLAlchemy version

# Google Services
google-auth-oauthlib==1.1.0
google-auth==2.23.3
google-api-python-client==2.108.0
google-auth-httplib2==0.1.0

# Utilities
python-dotenv==1.0.0
PyJWT==2.8.0
Werkzeug==2.3.7
python-dateutil==2.8.2
requests==2.31.0
Pillow==10.1.0
email-validator==2.1.0  # For email validation
python-slugify==8.0.1  # For URL-friendly strings
marshmallow==3.20.1  # For serialization/deserialization

# AI/ML
cohere==4.37

# Testing
pytest==8.0.2
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-flask==1.3.0  # For testing Flask applications
factory-boy==3.3.0  # For test data generation

# Production
gunicorn==21.2.0
sentry-sdk[flask]==1.39.1  # For error tracking
